827-refactor: Fsd core components (#848)
* refactor: 827 - move core components to widget

* refactor: 827 - fix imports

* refactor: 827 - delete unnecessary comments and add alias

---------
