+++ src/core/base-layout/base-layout.test.tsx
@@ -6,11 +6,11 @@ import { ROUTES } from '@/core/const';
 
 const mockUsePathname = vi.fn();
 
-vi.mock('./components/header', () => ({ Header: vi.fn(() => <div data-testid="mockHeader" />) }));
+vi.mock('@/widgets/header', () => ({ Header: vi.fn(() => <div data-testid="mockHeader" />) }));
 
-vi.mock('./components/partnered', () => ({ Partnered: vi.fn(() => <div data-testid="mockPartnered" />) }));
+vi.mock('@/widgets/partnered', () => ({ Partnered: vi.fn(() => <div data-testid="mockPartnered" />) }));
 
-vi.mock('./components/footer', () => ({ Footer: vi.fn(() => <div data-testid="mockFooter" />) }));
+vi.mock('@/widgets/footer', () => ({ Footer: vi.fn(() => <div data-testid="mockFooter" />) }));
 
 vi.mock('next/navigation', () => ({
   usePathname() {
