+++ src/widgets/footer/footer.scss
