+++ src/widgets/footer/ui/about-list.tsx
