+++ src/widgets/footer/ui/desktop-view.tsx
