+++ src/widgets/header/ui/nav-item/nav-item.module.scss
